Restructured code. New packages, new classes, refined encapsulation.
The model package is divided into the following subpackages:
package model.chess_set
	Board.java
	Piece.java
	PieceSet.java
	package model.chess_set.piecetypes
		[all subclasses of Piece]
package game
	Game.java
	Player.java
	Position.java

All game functionality is done through a Game object.
Instantiate a Game in the client like so:

Game game = new Game();

With this instance, Game can
whitePlayMove, blackPlayMove from the client.

getWhitePlayPosition, getBlackPlayPosition, getWhiteNewPosition
getBlackNewPosition is for other classes within Game only.
(for use with Player).

Game --> Player --> Board --> Piece.
It may seem like a lot of intermediary steps to move a piece,
but enforcing encapsulation was important to the design.

Certain classes cannot be instantiated from within Chess.java
(the client source file), for it is in a different package,
separated from the model package and its subpackages -- which was
the goal.

In short, everything having to do with the chess match is done within
the Game class. The Player is like a character -- it can do what it is
defined to do (within bounds), but the Board governs what will be
possible within the bounds of the Player's activities.
A given subclass of Piece has certain abilities, different to its other
siblings.

More functionality can be built from what is present,
I am thinking about employing the use of a Dictionary
for the PieceSet as opposed to an array.

It will making getting a Piece by its Position must more efficient.
NavigableMap<Position, Piece> pieceMap = new TreeMap<Position, Piece>().

As of now, getPieceByPosition within PieceSet iterates through the array
within PieceSet until it finds a Piece with the given Position param,
or returns null. It can be done more efficiently with a Map data
structure.

We'll see.

(Damn. I did it again. I might as well write "Dear Diary" at the top of
every commit from now on......)