Fixed logic on movePiece() lines 102-156
First, we check to see if a move is legal.
(added isMoveLegal abstract method to Piece)

If the move is legal,
we check to see if there is another Piece
to take out at newPosition.

If so, we set newPositionCell.piece to null,
removing the Cell's reference to whatever
piece was there.
(ally pieces are ignored)

If a piece was taken out successfully,
or there is no piece at newPosition:

We retrieve the Cell associated with
the requested Piece's position (we haven't
moved it yet!)

Then we set oldPositionCell.piece to null,
cutting that cell's reference to a piece.

Then we call piece.move(newPosition)
to actually move the piece.
This call changes the internal Position
field of a given Piece.

Finally, we set newPositionCell.piece = piece.
The Cell associated with newPosition now
has its Piece field associated with the Piece
we moved. This statement affects what is printed on the board during
board.toString().