Refactoring of Board.java (see body)
setPieceNullAtPosition() is a method
that was once formerly in Cell --
now that the subclasses of Piece
take a Board param instead of a Cell[][] param,
having this method in Board made more sense to me.

The local variable in movePiece() that was called
isLegalMove is now called pieceMoveLegal.

In movePiece(),
there is an integrity check to verify that the
move is "legitimate" to the current Board instance.

Integrity check:
- If white piece:
	- is piece.isWhite() true?
 	- is pieceSetRef a PieceType.Color.WHITE?
	- does pieceSetRef refer to whiteSet?
	- is piece the same Piece within pieceSetRef?
	- is piece the same Piece within whiteSet?
- If black piece:
	- is piece.isBlack() true?
	- is pieceSetRef a PieceType.Color.BLACK?
	- does pieceSetRef refer to blackSet?
 	- is piece the same Piece within pieceSetRef?
	- is piece the same Piece within blackSet?

getPieceSet() is a method that requires a Player
to be passed in. This ensures that the proper PieceSet is returned, and
that this method can really only be called where it is intended to be
called (in the Player class)

getPosition() is a method that returns the Position
object of a Cell, given a int file and int rank.

Instead of new Position objects being made each time something is to be
referred to by Position, I felt that it would be better to retrieve the
existing Position object that is associated with a Cell on the Board,
and use the returned Position object as a means to an end.

A new Position object should only be made for the instantiation of a
brand new Piece (like during a Promotion) -- but for now, I am still
assigning the promoter Piece the Position reference belonging to the
Pawn that it will be promoting. It may be better to use Position's copy
constructor to create a new Position object based off of an existing
Pawn on the board.

Anyway..

Move list reflects new addition to Piece -- Piece can now denote whether
it is 'alive' or 'dead'.

Also, class Move is no longer an inner class within Board.