Preparing for migration to GUI interface by making additions (see body)
-Methods in Game.java to retrieve
	last move
	last kill (all pieces know if they are alive or dead)

	Last kill will be handy because it will help me
	determine which graphical cell to remove the piece icon from.

-Accessor methods for all boolean fields in Game.java, to monitor
the state of the game
	We will no longer have a loop as per the CLI interface,
	so each time an piece icon is touched or clicked on the graphical
	board, things like isWhitesMove() or isValidMoveInput() will be
	checked for

-Bug fixes with Bishop
	A Bishop is a piece that moves diagonally (in an X pattern)
	either up 1, right 1 -- or down 1, left 1.

	It cannot hop over pieces like other pieces can,
	so if a player wants to move to a Cell that happens
	to be blocked by another piece, it cannot be allowed to move
	to that spot, or move past it.

	I have added a condition that accounts for either moving
	right, or moving left -- if moving right,
	the for-loop condition that allows it to continue is:
		x < pos.getFile()

	if moving left, the for-loop condition that allows it to
	continue is:
		x > pos.getFile()

	Saying x != pos.getFile() would cause an array-out-of bounds
	exception. Also -- once a Piece is found within the intended
	path of the Bishop (and the intended Position is along this path)
	-- result must be set false, and then break from the loop --
	since Bishops cannot hop over pieces.

The GUI interface will merely be a "handle" for all files
in the model package and subpackages --
so I have more work to do with Game.java to ensure smooth
migration to Android.

The tags generated by the android ImageView will be helpful,
as each square is denoted by <int><int><piecechar><piecechar> --
for example, a white pawn at a1 is noted as "00WP" --

So, if a user touches a pawn at a2 and wants to move it to a3 --
the corresponding ImageView tags are:

Old Position Cell/Piece	: "01WP"
New Position Cell		: "02--"		("--" means no piece at that cell)

I may translate something like "01WP02--" to "a1 a2", being that
such a string would already work as is with Game.java's methods --
or I may write an additional method in Game.java to use the tag
string as is.

Once Game.java/Board.java and company do its work as far as moving
a piece on the backend, I will call game.isValidMoveInput()
(plus the other methods...just simplifying this for cuz im tired)
to verify that the move made on the backend was good to go --
and if so, the graphical representation of the piece would follow suit.

More to come later tonight